requirements.txt
Kivy==2.3.0
kivy_garden.mapview==1.0.6
